# openapi.yaml
openapi: 3.1.0
info:
  title: Fusion E-Commerce Backend APIs
  description: |
    API documentation for the Fusion E-Commerce backend server. This documentation provides detailed information on all available endpoints for managing products, users, authentication, and more.
  version: 1.1.0
  contact:
    email: support@fusion-ecommerce.com
servers:
  - url: https://fusion-electronics-api.vercel.app/
    description: Production server
  - url: https://fusion-electronics-api.vercel.app
    description: Local development server
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Bad request
  /api/auth/login:
    post:
      summary: Authenticate a user and return a JWT token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials
  /api/auth/verify-email:
    post:
      summary: Verify if an email exists in the system
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Email verified
        '404':
          description: Email not found
  /api/auth/reset-password:
    post:
      summary: Reset a user's password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Password reset successfully
        '400':
          description: Bad request
  /api/checkout/create-order:
    post:
      summary: Create a new order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId:
                        type: string
                      quantity:
                        type: integer
                totalAmount:
                  type: number
      responses:
        '201':
          description: Order created successfully
        '400':
          description: Bad request
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string